+++ src/main/resources/templates/template_remand-and-sentencing-api.mustache
@@ -1,4 +1,4 @@
-<h1 class="title">Remand and Sentencing</h1>
+<h1 class="title">RaS-Recalls-Immigration-Detention</h1>
 {{#.}}
     <h2>Offender Details</h2>
     <table class="summary-list">
@@ -8,6 +8,7 @@
     <h3>Immigration Detentions</h3>
     {{#immigrationDetentions}}
         <table class="data-table">
+            <tr><td>Details</td></tr>
             <tr><td>Home Office Ref.</td><td>{{ optionalValue homeOfficeReferenceNumber }}</td></tr>
             <tr><td>No Longer of Interest Reason</td><td>{{ optionalValue noLongerOfInterestReason }}</td></tr>
             <tr><td>No Longer of Interest Comment</td><td>{{ optionalValue noLongerOfInterestComment }}</td></tr>
