--- src/test/resources/sar/sar-generated-report.html
@@ -1,27 +1,0 @@
-<html>
-<head>
-  <style>*{color:#0b0c0c;font-family:"HelveticaNeue","Helvetica Neue","Arial","Helvetica",sans-serif;font-size:12pt}.title{font-size:20pt;text-align:center}h1,h2,h3{color:#0b0c0c;display:block;font-weight:700;margin-inline-end:0;margin-inline-start:0;margin-top:0}h1{font-size:18pt;line-height:1.09375;margin-block-end:0.67em;margin-block-start:0.67em;margin-bottom:10px;margin-top:20px}h2{font-size:17pt;line-height:1.0416666667;margin-block-end:0.83em;margin-block-start:0.83em;margin-bottom:5px;margin-top:15px}h3{font-size:16pt;line-height:1.0416666667;margin-block-end:0.83em;margin-block-start:0.83em;margin-bottom:5px;margin-top:15px}h4{font-size:15pt;line-height:1.0416666667;margin-block-end:0.83em;margin-block-start:0.83em;margin-bottom:5px;margin-top:15px}h5{font-size:14pt;line-height:1.0416666667;margin-block-end:0.83em;margin-block-start:0.83em;margin-bottom:5px;margin-top:15px}table.summary-list{border-collapse:collapse;display:table;margin:0 0 5px;margin-block-end:1em;margin-block-start:1em;margin-inline-end:0;margin-inline-start:0;table-layout:fixed;width:700px}table.summary-list tr{display:table-row;margin:0 0 20px}table.summary-list tr td{border-bottom:1px solid #b1b4b6;border-collapse:collapse;display:table-cell;font-weight:400;line-height:1.25;margin-inline-start:40px;overflow-wrap:break-word;table-layout:fixed;vertical-align:top;white-space:normal;width:70%}table.summary-list tr td:first-child{font-size:12pt;font-weight:700;padding-right:20px;vertical-align:top;width:30%}table.data-table{border-collapse:collapse;display:table;margin:0 0 5px;margin-block-end:1em;margin-block-start:1em;margin-inline-end:0;margin-inline-start:0;table-layout:fixed;width:700px}table.data-table tr{display:table-row;margin:0 0 20px}table.data-table tr td{border-bottom:1px solid #b1b4b6;border-collapse:collapse;display:table-cell;font-weight:400;line-height:1.25;margin-bottom:5px;margin-inline-start:40px;overflow-wrap:break-word;padding-bottom:2px;padding-right:0;padding-top:2px;table-layout:fixed;vertical-align:top;white-space:pre-wrap}table.data-table tr:first-child td{border-bottom:none;font-size:12pt;font-weight:700;padding-right:20px;vertical-align:top;width:30%}table.data-table tr td.data-column-5{width:5%}table.data-table tr td.data-column-10{width:10%}table.data-table tr td.data-column-15{width:15%}table.data-table tr td.data-column-20{width:20%}table.data-table tr td.data-column-25{width:25%}table.data-table tr td.data-column-30{width:30%}table.data-table tr td.data-column-35{width:35%}table.data-table tr td.data-column-40{width:40%}table.data-table tr td.data-column-45{width:45%}table.data-table tr td.data-column-50{width:50%}table.data-table tr td.data-column-55{width:55%}table.data-table tr td.data-column-60{width:60%}table.data-table tr td.data-column-65{width:65%}table.data-table tr td.data-column-70{width:70%}table.data-table tr td.data-column-75{width:75%}table.data-table tr td.data-column-80{width:80%}table.data-table tr td.data-column-85{width:85%}table.data-table tr td.data-column-90{width:90%}table.data-table tr td.data-column-95{width:95%}table tr td.data-column-5{width:5%}table tr td.data-column-10{width:10%}table tr td.data-column-15{width:15%}table tr td.data-column-20{width:20%}table tr td.data-column-25{width:25%}table tr td.data-column-30{width:30%}table tr td.data-column-35{width:35%}table tr td.data-column-40{width:40%}table tr td.data-column-45{width:45%}table tr td.data-column-50{width:50%}table tr td.data-column-55{width:55%}table tr td.data-column-60{width:60%}table tr td.data-column-65{width:65%}table tr td.data-column-70{width:70%}table tr td.data-column-75{width:75%}table tr td.data-column-80{width:80%}table tr td.data-column-85{width:85%}table tr td.data-column-90{width:90%}table tr td.data-column-95{width:95%}table.summary-list tr td.data-column-5{width:5%}table.summary-list tr td.data-column-10{width:10%}table.summary-list tr td.data-column-15{width:15%}table.summary-list tr td.data-column-20{width:20%}table.summary-list tr td.data-column-25{width:25%}table.summary-list tr td.data-column-30{width:30%}table.summary-list tr td.data-column-35{width:35%}table.summary-list tr td.data-column-40{width:40%}table.summary-list tr td.data-column-45{width:45%}table.summary-list tr td.data-column-50{width:50%}table.summary-list tr td.data-column-55{width:55%}table.summary-list tr td.data-column-60{width:60%}table.summary-list tr td.data-column-65{width:65%}table.summary-list tr td.data-column-70{width:70%}table.summary-list tr td.data-column-75{width:75%}table.summary-list tr td.data-column-80{width:80%}table.summary-list tr td.data-column-85{width:85%}table.summary-list tr td.data-column-90{width:90%}table.summary-list tr td.data-column-95{width:95%}ul{padding:0 0 0 10px}</style>
-</head>
-<body>
-<h1 class="title">Remand and Sentencing</h1>
-
-<h2>Offender Details</h2>
-<table class="summary-list">
-  <tr><td>Prisoner Number</td><td>PRI123</td></tr>
-</table>
-
-<h3>Immigration Detentions</h3>
-
-<table class="data-table">
-  <tr><td>Home Office Ref.</td><td>3240593452</td></tr>
-  <tr><td>No Longer of Interest Reason</td><td>RIGHT_TO_REMAIN</td></tr>
-  <tr><td>No Longer of Interest Comment</td><td>Currently has the right to Remain</td></tr>
-</table>
-
-<table class="data-table">
-  <tr><td>Home Office Ref.</td><td>5340593452</td></tr>
-  <tr><td>No Longer of Interest Reason</td><td>BRITISH_CITIZEN</td></tr>
-  <tr><td>No Longer of Interest Comment</td><td>Recently made a British Citizen</td></tr>
-</table>
-</body>
-</html>
